@@ -11,15 +11,16 @@
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
   },
   "devDependencies": {
-    "@sveltejs/kit": "^2.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/kit": "^2.5.27",
     "prettier": "^3.2.5",
-    "prettier-plugin-svelte": "^3.2.3",
-    "svelte": "^4.2.9",
-    "svelte-check": "^3.6.3",
+    "prettier-plugin-svelte": "^3.2.6",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
     "tslib": "^2.6.2",
-    "typescript": "^5.3.3",
-    "vite": "^5.0.0"
+    "typescript": "^5.5.0",
+    "vite": "^5.4.4",
+    "@efstajas/svelte-stored-writable": "^0.2.0",
+    "@sveltejs/adapter-node": "^5.0.1"
   },
   "type": "module",
   "dependencies": {
